refactor: 커스텀 훅, 스토어 장소 관련 용어 content로 변경
