+++ components/Layout/Contents/ContentSearchInput.styles.ts
