+++ components/Layout/Contents/ContentSearchModal.styles.ts
